clean-up displacementestimation module
